+++ administrator/components/com_banners/Controller/Client.php
@@ -6,15 +6,24 @@
  * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
+namespace Joomla\Component\Banners\Administrator\Controller;
 
 defined('_JEXEC') or die;
 
+use Joomla\CMS\Controller\Form;
+
 /**
- * Banners component helper.
+ * Client controller class.
  *
  * @since  1.6
  */
-class BannersHelper extends \Joomla\Component\Banners\Administrator\Helper\BannersHelper
+class Client extends Form
 {
-
+	/**
+	 * The prefix to use with controller messages.
+	 *
+	 * @var    string
+	 * @since  1.6
+	 */
+	protected $text_prefix = 'COM_BANNERS_CLIENT';
 }
