+++ administrator/components/com_banners/Controller/Controller.php
@@ -1,0 +1,69 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_banners
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Banners\Administrator\Controller;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Controller\Controller as BaseController;
+use Joomla\Component\Banners\Administrator\Helper\BannersHelper;
+
+/**
+ * Banners master display controller.
+ *
+ * @since  1.6
+ */
+class Controller extends BaseController
+{
+	/**
+	 * The default view.
+	 *
+	 * @var    string
+	 * @since  1.6
+	 */
+	protected $default_view = 'banners';
+
+	/**
+	 * Method to display a view.
+	 *
+	 * @param   boolean  $cachable   If true, the view output will be cached
+	 * @param   array    $urlparams  An array of safe URL parameters and their variable types, for valid values see {@link \JFilterInput::clean()}.
+	 *
+	 * @return  static  This object to support chaining.
+	 *
+	 * @since   1.5
+	 */
+	public function display($cachable = false, $urlparams = array())
+	{
+		BannersHelper::updateReset();
+
+		$view   = $this->input->get('view', 'banners');
+		$layout = $this->input->get('layout', 'default');
+		$id     = $this->input->getInt('id');
+
+		// Check for edit form.
+		if ($view == 'banner' && $layout == 'edit' && !$this->checkEditId('com_banners.edit.banner', $id))
+		{
+			// Somehow the person just went to the form - we don't allow that.
+			$this->setMessage(\JText::sprintf('JLIB_APPLICATION_ERROR_UNHELD_ID', $id), 'error');
+			$this->setRedirect(\JRoute::_('index.php?option=com_banners&view=banners', false));
+
+			return false;
+		}
+		elseif ($view == 'client' && $layout == 'edit' && !$this->checkEditId('com_banners.edit.client', $id))
+		{
+			// Somehow the person just went to the form - we don't allow that.
+			$this->setMessage(\JText::sprintf('JLIB_APPLICATION_ERROR_UNHELD_ID', $id), 'error');
+			$this->setRedirect(\JRoute::_('index.php?option=com_banners&view=clients', false));
+
+			return false;
+		}
+
+		return parent::display();
+	}
+}
