+++ administrator/components/com_banners/Table/Client.php
@@ -1,0 +1,128 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_banners
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Banners\Administrator\Table;
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Table\Table;
+use Joomla\Utilities\ArrayHelper;
+
+/**
+ * Client table
+ *
+ * @since  1.6
+ */
+class Client extends Table
+{
+	/**
+	 * Constructor
+	 *
+	 * @param   \JDatabaseDriver  $db  Database connector object
+	 *
+	 * @since   1.5
+	 */
+	public function __construct(\JDatabaseDriver $db)
+	{
+		$this->typeAlias        = 'com_banners.client';
+		$this->checked_out_time = $db->getNullDate();
+
+		parent::__construct('#__banner_clients', 'id', $db);
+	}
+
+	/**
+	 * Method to set the publishing state for a row or list of rows in the database
+	 * table.  The method respects checked out rows by other users and will attempt
+	 * to checkin rows that it can after adjustments are made.
+	 *
+	 * @param   mixed    $pks     An optional array of primary key values to update.  If not set the instance property value is used.
+	 * @param   integer  $state   The publishing state. eg. [0 = unpublished, 1 = published, 2=archived, -2=trashed]
+	 * @param   integer  $userId  The user id of the user performing the operation.
+	 *
+	 * @return  boolean  True on success.
+	 *
+	 * @since   1.0.4
+	 */
+	public function publish($pks = null, $state = 1, $userId = 0)
+	{
+		$k = $this->_tbl_key;
+
+		// Sanitize input.
+		$pks    = ArrayHelper::toInteger($pks);
+		$userId = (int) $userId;
+		$state  = (int) $state;
+
+		// If there are no primary keys set check to see if the instance key is set.
+		if (empty($pks))
+		{
+			if ($this->$k)
+			{
+				$pks = array($this->$k);
+			}
+			// Nothing to set publishing state on, return false.
+			else
+			{
+				$this->setError(\JText::_('JLIB_DATABASE_ERROR_NO_ROWS_SELECTED'));
+
+				return false;
+			}
+		}
+
+		// Build the WHERE clause for the primary keys.
+		$where = $k . '=' . implode(' OR ' . $k . '=', $pks);
+
+		// Determine if there is checkin support for the table.
+		if (!property_exists($this, 'checked_out') || !property_exists($this, 'checked_out_time'))
+		{
+			$checkin = '';
+		}
+		else
+		{
+			$checkin = ' AND (checked_out = 0 OR checked_out = ' . (int) $userId . ')';
+		}
+
+		// Update the publishing state for rows with the given primary keys.
+		$this->_db->setQuery(
+			'UPDATE ' . $this->_db->quoteName($this->_tbl)
+			. ' SET ' . $this->_db->quoteName('state') . ' = ' . (int) $state
+			. ' WHERE (' . $where . ')'
+			. $checkin
+		);
+
+		try
+		{
+			$this->_db->execute();
+		}
+		catch (\RuntimeException $e)
+		{
+			$this->setError($e->getMessage());
+
+			return false;
+		}
+
+		// If checkin is supported and all rows were adjusted, check them in.
+		if ($checkin && (count($pks) == $this->_db->getAffectedRows()))
+		{
+			// Checkin the rows.
+			foreach ($pks as $pk)
+			{
+				$this->checkin($pk);
+			}
+		}
+
+		// If the \JTable instance value is in the list of primary keys that were set, set the instance.
+		if (in_array($this->$k, $pks))
+		{
+			$this->state = $state;
+		}
+
+		$this->setError('');
+
+		return true;
+	}
+}
