+++ administrator/components/com_banners/View/Banner/Html.php
@@ -1,0 +1,132 @@
+<?php
+/**
+ * @package     Joomla.Administrator
+ * @subpackage  com_banners
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Banners\Administrator\View\Banner;
+
+use Joomla\CMS\View\HtmlView;
+
+defined('_JEXEC') or die;
+
+
+/**
+ * View to edit a banner.
+ *
+ * @since  1.5
+ */
+class Html extends HtmlView
+{
+	/**
+	 * The \JForm object
+	 *
+	 * @var  \JForm
+	 */
+	protected $form;
+
+	/**
+	 * The active item
+	 *
+	 * @var  object
+	 */
+	protected $item;
+
+	/**
+	 * The model state
+	 *
+	 * @var  object
+	 */
+	protected $state;
+
+	/**
+	 * Display the view
+	 *
+	 * @param   string  $tpl  The name of the template file to parse; automatically searches through the template paths.
+	 *
+	 * @return  mixed  A string if successful, otherwise an Error object.
+	 */
+	public function display($tpl = null)
+	{
+		// Initialiase variables.
+		$this->form  = $this->get('Form');
+		$this->item  = $this->get('Item');
+		$this->state = $this->get('State');
+
+		// Check for errors.
+		if (count($errors = $this->get('Errors')))
+		{
+			throw new \JViewGenericdataexception(implode("\n", $errors), 500);
+		}
+
+		$this->addToolbar();
+
+		return parent::display($tpl);
+	}
+
+	/**
+	 * Add the page title and toolbar.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.6
+	 */
+	protected function addToolbar()
+	{
+		\JFactory::getApplication()->input->set('hidemainmenu', true);
+
+		$user       = \JFactory::getUser();
+		$userId     = $user->id;
+		$isNew      = ($this->item->id == 0);
+		$checkedOut = !($this->item->checked_out == 0 || $this->item->checked_out == $userId);
+
+		// Since we don't track these assets at the item level, use the category id.
+		$canDo = \JHelperContent::getActions('com_banners', 'category', $this->item->catid);
+
+		\JToolbarHelper::title($isNew ? \JText::_('COM_BANNERS_MANAGER_BANNER_NEW') : \JText::_('COM_BANNERS_MANAGER_BANNER_EDIT'), 'bookmark banners');
+
+		$toolbarButtons = [];
+
+		// If not checked out, can save the item.
+		if (!$checkedOut && ($canDo->get('core.edit') || count($user->getAuthorisedCategories('com_banners', 'core.create')) > 0))
+		{
+			$toolbarButtons[] = ['apply', 'banner.apply'];
+			$toolbarButtons[] = ['save', 'banner.save'];
+
+			if ($canDo->get('core.create'))
+			{
+				$toolbarButtons[] = ['save2new', 'banner.save2new'];
+			}
+		}
+
+		// If an existing item, can save to a copy.
+		if (!$isNew && $canDo->get('core.create'))
+		{
+			$toolbarButtons[] = ['save2copy', 'banner.save2copy'];
+		}
+
+		\JToolbarHelper::saveGroup(
+			$toolbarButtons,
+			'btn-success'
+		);
+
+		if (empty($this->item->id))
+		{
+			\JToolbarHelper::cancel('banner.cancel');
+		}
+		else
+		{
+			if (\JComponentHelper::isEnabled('com_contenthistory') && $this->state->params->get('save_history', 0) && $canDo->get('core.edit'))
+			{
+				\JToolbarHelper::versions('com_banners.banner', $this->item->id);
+			}
+
+			\JToolbarHelper::cancel('banner.cancel', 'JTOOLBAR_CLOSE');
+		}
+
+		\JToolbarHelper::divider();
+		\JToolbarHelper::help('JHELP_COMPONENTS_BANNERS_BANNERS_EDIT');
+	}
+}
