--- administrator/components/com_banners/banners.php
@@ -1,21 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_banners
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-JHtml::_('behavior.tabstate');
-
-if (!JFactory::getUser()->authorise('core.manage', 'com_banners'))
-{
-	throw new JAccessExceptionNotallowed(JText::_('JERROR_ALERTNOAUTHOR'), 403);
-}
-
-// Execute the task.
-$controller = JControllerLegacy::getInstance('Banners');
-$controller->execute(JFactory::getApplication()->input->get('task'));
-$controller->redirect();
