--- administrator/components/com_banners/controller.php
@@ -1,59 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_banners
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-JLoader::register('BannersHelper', JPATH_ADMINISTRATOR . '/components/com_banners/helpers/banners.php');
-
-/**
- * Banners master display controller.
- *
- * @since  1.6
- */
-class BannersController extends JControllerLegacy
-{
-	/**
-	 * Method to display a view.
-	 *
-	 * @param   boolean  $cachable   If true, the view output will be cached
-	 * @param   array    $urlparams  An array of safe URL parameters and their variable types, for valid values see {@link JFilterInput::clean()}.
-	 *
-	 * @return  BannersController  This object to support chaining.
-	 *
-	 * @since   1.5
-	 */
-	public function display($cachable = false, $urlparams = array())
-	{
-		BannersHelper::updateReset();
-
-		$view   = $this->input->get('view', 'banners');
-		$layout = $this->input->get('layout', 'default');
-		$id     = $this->input->getInt('id');
-
-		// Check for edit form.
-		if ($view == 'banner' && $layout == 'edit' && !$this->checkEditId('com_banners.edit.banner', $id))
-		{
-			// Somehow the person just went to the form - we don't allow that.
-			$this->setMessage(JText::sprintf('JLIB_APPLICATION_ERROR_UNHELD_ID', $id), 'error');
-			$this->setRedirect(JRoute::_('index.php?option=com_banners&view=banners', false));
-
-			return false;
-		}
-		elseif ($view == 'client' && $layout == 'edit' && !$this->checkEditId('com_banners.edit.client', $id))
-		{
-			// Somehow the person just went to the form - we don't allow that.
-			$this->setMessage(JText::sprintf('JLIB_APPLICATION_ERROR_UNHELD_ID', $id), 'error');
-			$this->setRedirect(JRoute::_('index.php?option=com_banners&view=clients', false));
-
-			return false;
-		}
-
-		return parent::display();
-	}
-}
