--- administrator/components/com_banners/controllers/client.php
@@ -1,26 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_banners
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-/**
- * Client controller class.
- *
- * @since  1.6
- */
-class BannersControllerClient extends JControllerForm
-{
-	/**
-	 * The prefix to use with controller messages.
-	 *
-	 * @var    string
-	 * @since  1.6
-	 */
-	protected $text_prefix = 'COM_BANNERS_CLIENT';
-}
