--- administrator/components/com_banners/views/client/view.html.php
@@ -1,137 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_banners
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-JLoader::register('BannersHelper', JPATH_ADMINISTRATOR . '/components/com_banners/helpers/banners.php');
-
-/**
- * View to edit a client.
- *
- * @since  1.5
- */
-class BannersViewClient extends JViewLegacy
-{
-	/**
-	 * The JForm object
-	 *
-	 * @var  JForm
-	 */
-	protected $form;
-
-	/**
-	 * The active item
-	 *
-	 * @var  object
-	 */
-	protected $item;
-
-	/**
-	 * The model state
-	 *
-	 * @var  object
-	 */
-	protected $state;
-
-	/**
-	 * Object containing permissions for the item
-	 *
-	 * @var  JObject
-	 */
-	protected $canDo;
-
-	/**
-	 * Display the view
-	 *
-	 * @param   string  $tpl  The name of the template file to parse; automatically searches through the template paths.
-	 *
-	 * @return  mixed  A string if successful, otherwise an Error object.
-	 */
-	public function display($tpl = null)
-	{
-		$this->form  = $this->get('Form');
-		$this->item  = $this->get('Item');
-		$this->state = $this->get('State');
-		$this->canDo = JHelperContent::getActions('com_banners');
-
-		// Check for errors.
-		if (count($errors = $this->get('Errors')))
-		{
-			throw new JViewGenericdataexception(implode("\n", $errors), 500);
-		}
-
-		$this->addToolbar();
-
-		return parent::display($tpl);
-	}
-
-	/**
-	 * Add the page title and toolbar.
-	 *
-	 * @return  void
-	 *
-	 * @since   1.6
-	 */
-	protected function addToolbar()
-	{
-		JFactory::getApplication()->input->set('hidemainmenu', true);
-
-		$user       = JFactory::getUser();
-		$isNew      = ($this->item->id == 0);
-		$checkedOut = !($this->item->checked_out == 0 || $this->item->checked_out == $user->id);
-		$canDo      = $this->canDo;
-
-		JToolbarHelper::title(
-			$isNew ? JText::_('COM_BANNERS_MANAGER_CLIENT_NEW') : JText::_('COM_BANNERS_MANAGER_CLIENT_EDIT'),
-			'bookmark banners-clients'
-		);
-
-		$toolbarButtons = [];
-
-		// If not checked out, can save the item.
-		if (!$checkedOut && ($canDo->get('core.edit') || $canDo->get('core.create')))
-		{
-			$toolbarButtons[] = ['apply', 'client.apply'];
-			$toolbarButtons[] = ['save', 'client.save'];
-		}
-
-		if (!$checkedOut && $canDo->get('core.create'))
-		{
-			$toolbarButtons[] = ['save2new', 'client.save2new'];
-		}
-
-		// If an existing item, can save to a copy.
-		if (!$isNew && $canDo->get('core.create'))
-		{
-			$toolbarButtons[] = ['save2copy', 'client.save2copy'];
-		}
-
-		JToolbarHelper::saveGroup(
-			$toolbarButtons,
-			'btn-success'
-		);
-
-		if (empty($this->item->id))
-		{
-			JToolbarHelper::cancel('client.cancel');
-		}
-		else
-		{
-			if (JComponentHelper::isEnabled('com_contenthistory') && $this->state->params->get('save_history', 0) && $canDo->get('core.edit'))
-			{
-				JToolbarHelper::versions('com_banners.client', $this->item->id);
-			}
-
-			JToolbarHelper::cancel('client.cancel', 'JTOOLBAR_CLOSE');
-		}
-
-		JToolbarHelper::divider();
-		JToolbarHelper::help('JHELP_COMPONENTS_BANNERS_CLIENTS_EDIT');
-	}
-}
