--- administrator/components/com_banners/views/download/view.html.php
@@ -1,45 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_banners
- *
- * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-/**
- * View class for download a list of tracks.
- *
- * @since  1.6
- */
-class BannersViewDownload extends JViewLegacy
-{
-	/**
-	 * The JForm object
-	 *
-	 * @var  JForm
-	 */
-	protected $form;
-
-	/**
-	 * Display the view
-	 *
-	 * @param   string  $tpl  The name of the template file to parse; automatically searches through the template paths.
-	 *
-	 * @return  mixed  A string if successful, otherwise an Error object.
-	 */
-	public function display($tpl = null)
-	{
-		$this->form = $this->get('Form');
-
-		// Check for errors.
-		if (count($errors = $this->get('Errors')))
-		{
-			throw new JViewGenericdataexception(implode("\n", $errors), 500);
-		}
-
-		return parent::display($tpl);
-	}
-}
