+++ components/com_banners/Controller/Controller.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * @package     Joomla.Site
+ * @subpackage  com_banners
+ *
+ * @copyright   Copyright (C) 2005 - 2017 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+namespace Joomla\Component\Banners\Site\Controller;
+
+use Joomla\CMS\Controller\Controller as BaseController;
+
+defined('_JEXEC') or die;
+
+/**
+ * Banners Controller
+ *
+ * @since  1.5
+ */
+class Controller extends BaseController
+{
+	/**
+	 * Method when a banner is clicked on.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.5
+	 */
+	public function click()
+	{
+		$id = $this->input->getInt('id', 0);
+
+		if ($id)
+		{
+			$model = $this->getModel('Banner', 'Site', array('ignore_request' => true));
+			$model->setState('banner.id', $id);
+			$model->click();
+			$this->setRedirect($model->getUrl());
+		}
+	}
+}
