+++ modules/mod_banners/helper.php
@@ -9,6 +9,8 @@
 
 defined('_JEXEC') or die;
 
+use Joomla\Component\Banners\Site\Model\Banners;
+
 /**
  * Helper for mod_banners
  *
@@ -25,13 +27,11 @@ class ModBannersHelper
 	 */
 	public static function &getList(&$params)
 	{
-		JModelLegacy::addIncludePath(JPATH_ROOT . '/components/com_banners/models', 'BannersModel');
-
 		$document = JFactory::getDocument();
 		$app      = JFactory::getApplication();
 		$keywords = explode(',', $document->getMetaData('keywords'));
 
-		$model = JModelLegacy::getInstance('Banners', 'BannersModel', array('ignore_request' => true));
+		$model = new Banners(array('ignore_request' => true));
 		$model->setState('filter.client_id', (int) $params->get('cid'));
 		$model->setState('filter.category_id', $params->get('catid', array()));
 		$model->setState('list.limit', (int) $params->get('count', 1));
